projects/lesson2_memory_trading/requirements.txt
langchain==0.1.0
openai
python-dotenv
tiktoken